(wip) daemon, o/h/ctlcmd: move state to client change conversion to ctlcmd
